// src/types/configs/enums.ts
export enum Roles {
	Admin = "admin",
	Editor = "editor",
	Author = "author",
	Producer = "producer",
	User = "user",
}

export enum ProductStatus {
	Inprogress = "Inprogress",
	Accepted = "Accepted",
	Rejected = "Rejected",
	TopProduct = "TopProduct",
}

export enum ProductLevel {
	Potential = "Potential",
	Augmented = "Augmented",
	Expected = "Expected",
	Generic = "Generic",
	Core = "Core",
}

export enum ProducerStatus {
	NotRequested = "NotRequested",
	Inprogress = "Inprogress",
	Accepted = "Accepted",
	Rejected = "Rejected",
	TopProducer = "TopProducer",
}